# specs/001-halal-stock-screener/contracts/api.yaml
openapi: 3.0.3
info:
  title: Halal TSX Stock Screener API
  description: API for viewing, filtering, and analyzing Halal-compliant stocks from the Toronto Stock Exchange
  version: 1.0.0
  contact:
    name: HalalTSX
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.halaltsx.com/v1
    description: Production server
...
paths:
  /stocks:
    get:
      tags:
        - Stocks
      summary: List all stocks
      description: Returns a paginated list of TSX stocks with basic information and compliance status
      operationId: getStocks
      parameters:
        - name: page
          in: query
          description: Page number (0-indexed)
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: search
          in: query
          description: Search by stock name or ticker symbol
          schema:
            type: string
        - name: compliance
          in: query
          description: Filter by compliance status
          schema:
            type: string
            enum: [ALL, COMPLIANT, NON_COMPLIANT, UNABLE_TO_VERIFY]
            default: ALL
        - name: sector
          in: query
          description: Filter by market sector
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, symbol, price, marketCap, sector]
            default: name
        - name: direction
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockListResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stocks/{symbol}:
    get:
      tags:
        - Stocks
      summary: Get stock details
      description: Returns detailed information about a specific stock including price history
      operationId: getStockBySymbol
      parameters:
        - name: symbol
          in: path
          required: true
          description: Stock ticker symbol (e.g., "RY.TO")
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockDetailResponse'
        '404':
          description: Stock not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    StockSummary:
      type: object
      required:
        - id
        - symbol
        - name
        - currentPrice
        - complianceStatus
        - priceUpdatedAt
      properties:
        id:
          type: integer
          format: int64
          description: Unique stock identifier
        symbol:
          type: string
          description: TSX ticker symbol
          example: "RY.TO"
        name:
          type: string
          description: Company name
          example: "Royal Bank of Canada"
        sector:
          type: string
          description: Market sector
          example: "Financials"
        currentPrice:
          type: number
          format: decimal
          description: Current stock price in CAD
          example: 145.23
        priceChange:
          type: number
          format: decimal
          description: Price change from previous close
          example: 2.15
        priceChangePercent:
          type: number
          format: decimal
          description: Percentage price change
          example: 1.50
        marketCap:
          type: number
          format: decimal
          description: Market capitalization in CAD
          example: 203500000000
        complianceStatus:
          type: string
          enum: [COMPLIANT, NON_COMPLIANT, UNABLE_TO_VERIFY]
          description: Overall Halal compliance status
        requiresPurification:
          type: boolean
          description: Whether dividend purification is required
        priceUpdatedAt:
          type: string
          format: date-time
          description: When price was last updated
    StockDetail:
      allOf:
        - $ref: '#/components/schemas/StockSummary'
        - type: object
          properties:
            subIndustry:
              type: string
              description: GICS sub-industry classification
            compliance:
              $ref: '#/components/schemas/ComplianceSummary'
    ComplianceSummary:
      type: object
      required:
        - status
        - screenedAt
      properties:
        status:
          type: string
          enum: [COMPLIANT, NON_COMPLIANT, UNABLE_TO_VERIFY]
        businessActivityCompliant:
          type: boolean
        debtRatioCompliant:
          type: boolean
        liquidityRatioCompliant:
          type: boolean
        incomeRatioCompliant:
          type: boolean
        requiresPurification:
          type: boolean
        purificationPercentage:
          type: number
          format: decimal
        screenedAt:
          type: string
          format: date-time
...
    StockListResponse:
      type: object
      required:
        - content
        - page
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/StockSummary'
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Items per page
        totalElements:
          type: integer
          format: int64
          description: Total number of stocks
        totalPages:
          type: integer
          description: Total number of pages
        dataFreshness:
          $ref: '#/components/schemas/DataFreshness'
    StockDetailResponse:
      type: object
      required:
        - stock
      properties:
        stock:
          $ref: '#/components/schemas/StockDetail'
        dataFreshness:
          $ref: '#/components/schemas/DataFreshness'
...
    DataFreshness:
      type: object
      properties:
        pricesUpdatedAt:
          type: string
          format: date-time
        financialsUpdatedAt:
          type: string
          format: date-time
        isStale:
          type: boolean
          description: True if data is older than acceptable threshold
        message:
          type: string
          description: Human-readable freshness status
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time